[IMP] (website_)payment, sale: move onboarding logic to res.company
The onboarding logic was originally implemented in `res.config.settings`, but
this transient model does not support all onboarding entry points. Specifically,
when the 'Activate <provider>' button was used in frontend payment forms, such
as the portal or checkout, the transient settings record could not be relied
upon. To work around this limitation, a temporary settings record had to be
generated from a QWeb template, which added unnecessary complexity and risk
of errors.

To resolve this issue, the onboarding logic is moved to `res.company`. The
company record is persistent and accessible in both backend and frontend
contexts. As a result, the 'Activate <provider>' button now reliably triggers
the onboarding process in all scenarios, without requiring temporary records.

task-5081953
