+++ addons/payment/data/ir_actions_server_data.xml
@@ -6,7 +6,7 @@
         <field name="model_id" ref="payment.model_payment_provider"/>
         <field name="state">code</field>
         <field name="code">
-action = env['res.config.settings'].sudo().create({})._start_payment_onboarding()
+action = env.company._start_payment_onboarding()
         </field>
     </record>
