+++ addons/payment/wizards/res_config_settings.py
@@ -15,15 +15,7 @@ class ResConfigSettings(models.TransientModel):
     has_enabled_provider = fields.Boolean(
         string="Has Enabled Provider", compute='_compute_has_enabled_provider'
     )
-    onboarding_payment_module = fields.Selection(
-        string="Onboarding Payment Module",
-        selection=[
-            ('mercado_pago', "Mercado Pago"),
-            ('razorpay', "Razorpay"),
-            ('stripe', "Stripe"),
-        ],
-        compute='_compute_onboarding_payment_module',
-    )
+    onboarding_payment_module = fields.Selection(related='company_id.onboarding_payment_module')
 
     # === COMPUTE METHODS === #
 
@@ -59,18 +51,6 @@ class ResConfigSettings(models.TransientModel):
         else:
             return company_domain & Domain('state', '!=', 'disabled')
 
-    @api.depends('company_id.currency_id', 'company_id.country_id.is_stripe_supported_country')
-    def _compute_onboarding_payment_module(self):
-        for config in self:
-            if config.company_id.currency_id.name == 'INR':
-                config.onboarding_payment_module = 'razorpay'
-            elif config.company_id.country_id.is_stripe_supported_country:
-                config.onboarding_payment_module = 'stripe'
-            elif config.company_id.country_id.is_mercado_pago_supported_country:
-                config.onboarding_payment_module = 'mercado_pago'
-            else:
-                config.onboarding_payment_module = None
-
     # === ACTION METHODS === #
 
     def action_view_active_provider(self):
@@ -82,33 +62,3 @@ class ResConfigSettings(models.TransientModel):
             'views': [[False, 'form']],
             'res_id': provider.id,
         }
-
-    def _start_payment_onboarding(self, menu_id=None):
-        """Install the onboarding module, configure the provider and run the onboarding action.
-
-        :param int menu_id: The menu from which the onboarding is started, as an `ir.ui.menu` id.
-        :return: The action returned by `action_start_onboarding`.
-        :rtype: dict
-        """
-        self.ensure_one()
-        if not self.onboarding_payment_module:
-            return False
-
-        # Install the onboarding module if needed.
-        onboarding_module = self.env['ir.module.module'].search(
-            [('name', '=', f'payment_{self.onboarding_payment_module}')]
-        )
-        self._install_modules(onboarding_module)
-        # Create a new env including the freshly installed module.
-        new_env = api.Environment(self.env.cr, self.env.uid, self.env.context)
-
-        # Configure the provider.
-        provider_code = self.onboarding_payment_module
-        provider = new_env['payment.provider'].search([
-            ('code', '=', provider_code),
-            *self.env['payment.provider']._check_company_domain(self.env.company),
-        ], limit=1)
-        if not provider:
-            return False
-
-        return provider.action_start_onboarding(menu_id=menu_id)
