+++ addons/website_payment/models/res_config_settings.py
@@ -30,4 +30,4 @@ class ResConfigSettings(models.TransientModel):
     # Unique name to avoid colliding with `sale`.
     def action_w_payment_start_payment_onboarding(self):
         menu = self.env.ref('website.menu_website_website_settings', raise_if_not_found=False)
-        return self._start_payment_onboarding(menu and menu.id)
+        return self.company_id._start_payment_onboarding(menu and menu.id)
